Optimize the code
